Rename `eln-validator` to `eln-archive`
This commit will be squashed into the next commit.
